move Hash and PubKey to their own module
